fix(docker-engine): merged volume files to match project structure
