--- yoki-engine-docker/src/commonMain/kotlin/org/katan/yoki/engine/docker/resource/volume/VolumePruneResponse.kt
@@ -1,11 +1,0 @@
-package org.katan.yoki.engine.docker.resource.volume
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-import org.katan.yoki.engine.docker.model.volume.Volume
-
-@Serializable
-public data class VolumePruneResponse(
-    @SerialName("VolumesDeleted") public val volumesDeleted: Collection<String>,
-    @SerialName("SpaceReclaimed") public val spaceReclaimed: Long
-)
